Move all Entity Viewer into it's own dir
