+++ src/UX/Entity/Detail/OutpointDetail.tsx
