+++ src/UX/Entity/Detail/UTXODetail.css
